requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
fastapi>=0.100.0
uvicorn>=0.23.0
ffmpeg-python>=0.2.0
